Configured proxysql

+++ scripts/proxysql.sh
@@ -10,7 +10,13 @@ if [ ! -f "${CONFIG_FILE_TEMPLATE}" ]; then
 fi
 
 printf "Generating custom config from template. "
-cat ${CONFIG_FILE_TEMPLATE} | envsubst > ${CONFIG_FILE}
-echo "Done."
+if [ ! -e "${CONFIG_FILE}.lock" ]; then
+    cat ${CONFIG_FILE_TEMPLATE} | envsubst > ${CONFIG_FILE}
+    touch ${CONFIG_FILE}.lock
+    echo "Done."
+else
+    echo "Skipped. Config file is already generated."
+    echo "For recreating config file from template, please recreate the container."
+fi
 
 exec $@
